# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ajloun Vision Trend Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /initiatives/{id}/volunteer:
    post:
      operationId: volunteerForInitiative
      tags: [volunteer]
      summary: Volunteer for an initiative
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Signed up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolunteerSignup"

  # ─── MEMBERS ───────────────────────────────────────────
  /members:
    get:
      operationId: listMembers
      tags: [members]
      parameters:
        - name: role
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: createMember
      tags: [members]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "404":
          description: Not found
    patch:
      operationId: updateMember
      tags: [members]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    delete:
      operationId: deleteMember
      tags: [members]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /members/{id}/impact:
    get:
      operationId: getMemberImpact
      tags: [members]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Impact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberImpact"

  # ─── KNOWLEDGE ─────────────────────────────────────────
...
  /opportunities:
    get:
      operationId: listOpportunities
      tags: [opportunities]
      parameters:
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Opportunities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Opportunity"
    post:
      operationId: createOpportunity
      tags: [opportunities]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpportunityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
  /opportunities/{id}:
    get:
      operationId: getOpportunity
      tags: [opportunities]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
    patch:
      operationId: updateOpportunity
      tags: [opportunities]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpportunityInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
    delete:
      operationId: deleteOpportunity
      tags: [opportunities]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── POINTS ────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
    post:
      operationId: sendNotification
      tags: [notifications]
      summary: Send notification to user(s) (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
...
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    # ─── AUTH ─────────────────────────────
...
    Initiative:
      type: object
      required: [id, title, titleAr, description, descriptionAr, status, category, participantsCount, isFeatured, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        status:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        participantsCount:
          type: integer
        targetParticipants:
          type: ["integer", "null"]
        progressPercent:
          type: ["integer", "null"]
        isFeatured:
          type: boolean
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        createdAt:
          type: string
...
    VolunteerSignup:
      type: object
      required: [id, userId, initiativeId, status, hoursLogged, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        initiativeId:
          type: integer
        status:
          type: string
        hoursLogged:
          type: integer
        notes:
          type: ["string", "null"]
        createdAt:
          type: string

    # ─── MEMBERS ──────────────────────────
    Member:
      type: object
      required: [id, fullName, fullNameAr, email, role, joinedAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        fullNameAr:
          type: string
        email:
          type: string
        role:
          type: string
        bio:
          type: ["string", "null"]
        bioAr:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        initiativesCount:
          type: integer
        impactScore:
          type: integer
        isActive:
          type: boolean
        joinedAt:
          type: string
    MemberInput:
      type: object
      required: [fullName, fullNameAr, email]
      properties:
        fullName:
          type: string
        fullNameAr:
          type: string
        email:
          type: string
        role:
          type: string
        bio:
          type: ["string", "null"]
        bioAr:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
    MemberUpdate:
      type: object
      properties:
        fullName:
          type: string
        fullNameAr:
          type: string
        email:
          type: string
        role:
          type: string
        bio:
          type: ["string", "null"]
        bioAr:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        isActive:
          type: boolean
    MemberImpact:
      type: object
      required: [memberId, initiativesJoined, initiativesLed, impactScore, badges]
      properties:
        memberId:
          type: integer
        initiativesJoined:
          type: integer
        initiativesLed:
          type: integer
        impactScore:
          type: integer
        badges:
          type: array
          items:
            type: string
        recentInitiatives:
          type: array
          items:
            $ref: "#/components/schemas/Initiative"

    # ─── KNOWLEDGE ────────────────────────
...
    Opportunity:
      type: object
      required: [id, title, titleAr, description, descriptionAr, type, organization, isRemote, isActive, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        type:
          type: string
        organization:
          type: string
        location:
          type: ["string", "null"]
        isRemote:
          type: boolean
        deadline:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    OpportunityInput:
      type: object
      required: [title, titleAr, description, descriptionAr, type, organization]
      properties:
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        type:
          type: string
        organization:
          type: string
        location:
          type: ["string", "null"]
        isRemote:
          type: boolean
        deadline:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
        isActive:
          type: boolean

    # ─── POINTS ───────────────────────────
...
    Notification:
      type: object
      required: [id, userId, type, title, titleAr, body, bodyAr, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        title:
          type: string
        titleAr:
          type: string
        body:
          type: string
        bodyAr:
          type: string
        isRead:
          type: boolean
        link:
          type: ["string", "null"]
        createdAt:
          type: string
    NotificationInput:
      type: object
      required: [titleAr, bodyAr]
      properties:
        userId:
          type: ["integer", "null"]
        type:
          type: string
        title:
          type: string
        titleAr:
          type: string
        body:
          type: string
        bodyAr:
          type: string
        link:
          type: ["string", "null"]
        sendToAll:
          type: boolean

    # ─── ADMIN ────────────────────────────
...