# lib/api-zod/src/generated/types/memberUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ajloun Vision Trend Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface MemberUpdate {
  fullName?: string;
  fullNameAr?: string;
  email?: string;
  role?: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  bioAr?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  location?: string | null;
  isActive?: boolean;
}